--- a/crates/lunatic-distributed/src/distributed/server.rs
+++ b/crates/lunatic-distributed/src/distributed/server.rs
@@ -19,22 +19,12 @@ use crate::{
 
 use super::message::Spawn;
 
-pub struct ServerCtx<T> {
+#[derive(Clone)]
+pub struct ServerCtx {
     pub envs: Environments,
-    pub modules: Modules<T>,
+    pub modules: Modules,
     pub distributed: DistributedProcessState,
     pub runtime: WasmtimeRuntime,
-}
-
-impl<T: 'static> Clone for ServerCtx<T> {
-    fn clone(&self) -> Self {
-        Self {
-            envs: self.envs.clone(),
-            modules: self.modules.clone(),
-            distributed: self.distributed.clone(),
-            runtime: self.runtime.clone(),
-        }
-    }
 }
 
 pub fn root_cert(test_ca: bool, ca_cert: Option<&str>) -> Result<String> {
@@ -59,7 +49,7 @@ pub fn gen_node_cert(node_name: &str) -> Result<Certificate> {
 }
 
 pub async fn node_server<T>(
-    ctx: ServerCtx<T>,
+    ctx: ServerCtx,
     socket: SocketAddr,
     cert: String,
     key: String,
@@ -68,31 +58,31 @@ where
     T: ProcessState + ResourceLimiter + DistributedCtx + Send + 'static,
 {
     let mut quic_server = quic::new_quic_server(socket, &cert, &key)?;
-    quic::handle_node_server(&mut quic_server, ctx.clone()).await?;
+    quic::handle_node_server::<T>(&mut quic_server, ctx.clone()).await?;
     Ok(())
 }
 
-pub async fn handle_message<T>(ctx: ServerCtx<T>, conn: quic::Connection, msg_id: u64, msg: Request)
+pub async fn handle_message<T>(ctx: ServerCtx, conn: quic::Connection, msg_id: u64, msg: Request)
 where
-    T: ProcessState + DistributedCtx + ResourceLimiter + Send + 'static,
+    T: ProcessState + ResourceLimiter + DistributedCtx + Send + 'static,
 {
-    if let Err(e) = handle_message_err(ctx, conn, msg_id, msg).await {
+    if let Err(e) = handle_message_err::<T>(ctx, conn, msg_id, msg).await {
         log::error!("Error handling message: {e}");
     }
 }
 
 async fn handle_message_err<T>(
-    ctx: ServerCtx<T>,
+    ctx: ServerCtx,
     conn: quic::Connection,
     msg_id: u64,
     msg: Request,
 ) -> Result<()>
 where
-    T: ProcessState + DistributedCtx + ResourceLimiter + Send + 'static,
+    T: ProcessState + ResourceLimiter + DistributedCtx + Send + 'static,
 {
     match msg {
         Request::Spawn(spawn) => {
-            let id = handle_spawn(ctx, spawn).await?;
+            let id = handle_spawn::<T>(ctx, spawn).await?;
             conn.send(msg_id, Response::Spawned(id)).await?;
         }
         Request::Message {
@@ -105,9 +95,9 @@ where
     Ok(())
 }
 
-async fn handle_spawn<T>(mut ctx: ServerCtx<T>, spawn: Spawn) -> Result<u64>
+async fn handle_spawn<T>(mut ctx: ServerCtx, spawn: Spawn) -> Result<u64>
 where
-    T: ProcessState + DistributedCtx + ResourceLimiter + Send + 'static,
+    T: ProcessState + ResourceLimiter + DistributedCtx + Send + 'static,
 {
     let Spawn {
         environment_id,
@@ -125,7 +115,9 @@ where
         None => {
             if let Some(bytes) = ctx.distributed.control.get_module(module_id).await {
                 let wasm = RawWasm::new(Some(module_id), bytes);
-                ctx.modules.compile(ctx.runtime.clone(), wasm).await??
+                ctx.modules
+                    .compile::<T>(ctx.runtime.clone(), wasm)
+                    .await??
             } else {
                 return Err(anyhow!("Cannot get the module from control"));
             }
@@ -138,21 +130,18 @@ where
     let state = T::new_dist_state(env.clone(), distributed, runtime, module.clone(), config)?;
     let params: Vec<wasmtime::Val> = params.into_iter().map(Into::into).collect();
     let (_handle, proc) = env
-        .spawn_wasm(ctx.runtime, module, state, &function, params, None)
+        .spawn_wasm::<T>(ctx.runtime, &module, state, &function, params, None)
         .await?;
     Ok(proc.id())
 }
 
-async fn handle_process_message<T>(
-    mut ctx: ServerCtx<T>,
+async fn handle_process_message(
+    mut ctx: ServerCtx,
     environment_id: u64,
     process_id: u64,
     tag: Option<i64>,
     data: Vec<u8>,
-) -> Result<()>
-where
-    T: ProcessState + DistributedCtx + ResourceLimiter + Send + 'static,
-{
+) -> Result<()> {
     let env = ctx.envs.get_or_create(environment_id);
     if let Some(proc) = env.get_process(process_id) {
         proc.send(Signal::Message(Message::Data(DataMessage::new_from_vec(