--- a/crates/lunatic-distributed/src/lib.rs
+++ b/crates/lunatic-distributed/src/lib.rs
@@ -15,7 +15,7 @@ pub trait DistributedCtx: ProcessState + Sized {
         environment: Environment,
         distributed: DistributedProcessState,
         runtime: WasmtimeRuntime,
-        module: WasmtimeCompiledModule<Self>,
+        module: Arc<WasmtimeCompiledModule>,
         config: Arc<Self::Config>,
     ) -> Result<Self>;
     fn distributed(&self) -> Result<&DistributedProcessState>;