Allows compiled wasm modules to be shared between the same node.

Two new host functions were added:
- `"lunatic::message::push_module"`
- `"lunatic::message::take_module"`

Modules are no longer pre-linked. The upside of this is that `WasmtimeCompiledModule` no longer has a generic (which removes generics from many other places too), but the downside is that it might be a little less performant.

I'll try to get some benchmarks done to test this PR's performance compared to before.