@tqwewe 2022-08-19
#### Main branch

```
spawn process           time:   [107.77 µs 108.37 µs 109.00 µs]                          
                        change: [-71.703% -71.432% -71.123%] (p = 0.00 < 0.05)
                        Performance has improved.
```

#### This PR

```
spawn process           time:   [384.67 µs 385.86 µs 387.08 µs]                          
                        change: [+251.58% +255.12% +258.40%] (p = 0.00 < 0.05)
                        Performance has regressed.
```

Seems like the performance does decrease substantially sadly :(

I'm not sure how to get past this since with pre linking requires `WasmtimeCompiledModule` to have a generic, which causes issues when sharing the module across processes.