@bkolobara 2022-08-19
We only have one state (generic `T`) though. Could we bubble it up all the way to `Resource<T>` and then use the state type to create the specific `Resource`? I didn't look at the code for some time now and am just assuming bunch of stuff.