@tqwewe 2022-08-29
After re-reading your comment, I think I understand.
The problem with `Resource` having a generic `T` is that we put `Resource` in a vec to store all resources.
```rust
pub struct DataMessage {
    // ...
    pub resources: Vec<Resource>,
}
```
If `Resource` had a generic here, then we couldnt store a vec of resources with different types.