@tqwewe 2022-08-29
The main requirement of the state is that it implements `ProcessState`. It seems like there's really only one type that implements `ProcessState`, and that is `DefaultProcessState`.
Could we remove `ProcessState` trait and rename `DefaultProcessState` to `ProcessState`? So `ProcessState` would be a sturct instead of a trait.