@bkolobara 2022-08-31
> The main requirement of the state is that it implements `ProcessState`. It seems like there's really only one type that implements `ProcessState`, and that is `DefaultProcessState`.
> Could we remove `ProcessState` trait and rename `DefaultProcessState` to `ProcessState`? So `ProcessState` would be a sturct instead of a trait.

The reason why we introduced the `ProcessState` trait is that in a case where we embed the lunatic_runtime inside of another rust project, you can provide your own process state and don't need to use all the features.

For example, if you don't need networking in your embedded usage, you just provide a `ProcessState` without the networking parts and don't implement `NetworkingCtx` for it. So you can still use all the other built in host functions that we provide, except the networking ones.

At the moment there is only one implementation of the `ProcessState` (`DefaultProcessState`), but this could change in the future. Even for "per feature (networking, timeouts, spawning, ...) testing purposes", we could provide smaller process states that are crate local. We would not be able to use the `DefaultProcessState` because of circular dependencies in that case.